#fix clean code

+++ packages/include/input/Event_MouseMotion.hpp
@@ -4,12 +4,12 @@
 namespace firefly::input {
 class Event_MouseMotion : public runtime::Event {
 private:
-  glm::vec2 _position;
-  glm::vec2 _delta;
+  glm::ivec2 _position;
+  glm::ivec2 _delta;
 
 public:
-  Event_MouseMotion(const glm::vec2 &position, const glm::vec2 &delta);
-  const glm::vec2 &getPosition() const;
-  const glm::vec2 &getDelta() const;
+  Event_MouseMotion(const glm::ivec2 &position, const glm::ivec2 &delta);
+  const glm::ivec2 &getPosition() const;
+  const glm::ivec2 &getDelta() const;
 };
 }; // namespace firefly::input
