+++ packages/include/runtime/Resource.hpp
@@ -5,7 +5,7 @@
 namespace firefly::runtime {
 class Resource : public core::Object {
 public:
-  virtual core::AutoPtr<core::Buffer> read(const size_t &size = 0) = 0;
+  virtual core::AutoPtr<core::Buffer> read(const uint32_t &size = 0) = 0;
   virtual void write(const core::AutoPtr<core::Buffer> &data) = 0;
 };
 } // namespace firefly::runtime
